Finalizando detalhes do projeto
