+++ labeddit/src/pages/PostPage/index.js
@@ -1,16 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import styled from 'styled-components';
 import Comments from './Comments';
 import { makeStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
-import Button from '@material-ui/core/Button';
 import { useParams } from 'react-router-dom';
 import { useHistory } from 'react-router-dom';
 import { useForm } from '../../hooks/useForm';
 import Header from '../../components/Header'
-
-
+import { PostPageContainer, Post, Title, User, Text, FormContainer, CommentsContainer, SendButton } from './PostPageStyles';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -21,50 +18,6 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-const PostPageContainer = styled.div`
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-`
-const Post = styled.div`
-    width: 40rem;
-    
-`
-const Title = styled.h2`
-    font-family: 'Special Elite', cursive;
-    font-size: 3rem;
-    margin: 1rem 0;
-`
-const User = styled.h3`
-    font-size: 1.5rem;
-    margin-bottom: 1rem;
-`
-const Text = styled.p`
-    font-size: 1.3rem;
-`
-
-const FormContainer = styled.form`
-    display: flex;
-    flex-direction: column;
-`
-const CommentsContainer = styled.div`
-    width: 40rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-`
-const SendButton = styled(Button)`
-    background-color: #ff5722;
-    color: white;
-    :hover {
-        color: #ff5722;
-        background-color: white;
-        border: 1px solid #ff5722;
-    }
-`
-
 const PostPage = () => {
     const classes = useStyles();
     const history = useHistory();
@@ -73,7 +26,6 @@ const PostPage = () => {
     const [comentarios, setComentarios] = useState([])
     const { form, onChange } = useForm({comment: ''})
 
-    
 
     const postDetails = async () => {
         try {
@@ -91,8 +43,7 @@ const PostPage = () => {
 
     useEffect(() => {
         postDetails();
-        
-    }, [])
+    })
 
     const handleInputChange = event => {
         const { name, value } = event.target;
@@ -136,7 +87,7 @@ const PostPage = () => {
             
             <CommentsContainer>
                 <p>{comentarios.length} Comentários:</p>
-                <Comments comments={comentarios}/>
+                <Comments comments={comentarios} id={pathParams.idPost}/>
             </CommentsContainer>
             
         </PostPageContainer>
